Split trash can into two assets
When the trash can was one art asset, pizzas were disappearing behind it (making it seem like we were littering) so I split the back wall and the front of the trash can into two art assets. When layered on top of one another, they look like a complete trash can, we just need to render the back before the pizzas, and the front after.